tweaking cubit logix

+++ lib/front/home_screen/cubit/transaction_cubit.dart
@@ -3,20 +3,15 @@ import 'package:equatable/equatable.dart';
 import 'package:pvp_projektas/backend/models/transaction.dart';
 import 'package:pvp_projektas/backend/transaction_repository/transaction_repository.dart';
 
-class TransactionState extends Equatable{
+class TransactionState extends Equatable {
   final List<Transaction> transactions;
 
-
-  const TransactionState({
-     this.transactions = const []
-  });
+  const TransactionState({this.transactions = const []});
 
   @override
   List<Object> get props => [transactions];
 
-  TransactionState copyWith({
-    List<Transaction>? transactions
-  }) =>
+  TransactionState copyWith({List<Transaction>? transactions}) =>
       TransactionState(
         transactions: transactions ?? this.transactions,
       );
@@ -25,13 +20,23 @@ class TransactionState extends Equatable{
 class TransactionCubit extends Cubit<TransactionState> {
   final TransactionRepository transRepository;
 
-  TransactionCubit(this.transRepository) : super(const TransactionState(transactions: [])){
+  TransactionCubit(this.transRepository)
+      : super(const TransactionState(transactions: [])) {
     loadTransactions();
   }
 
   void loadTransactions() {
     final transactions = transRepository.getTransactions();
     emit(TransactionState(transactions: transactions));
+  }
+
+  void saveTransaction(Transaction transaction, int? index) {
+    if(index != null){
+      updateTransaction(transaction, index);
+    }
+    else{
+      addTransaction(transaction);
+    }
   }
 
   void addTransaction(Transaction transaction) {
@@ -43,8 +48,8 @@ class TransactionCubit extends Cubit<TransactionState> {
     emit(state.copyWith(transactions: transactionsList));
   }
 
-  void updateTransaction(Transaction? transaction, int index) {
-    if(transaction != null){
+  void updateTransaction(Transaction transaction, int index) {
+    if (index >= 0 && index < state.transactions.length) {
       //create local copy of transactions
       List<Transaction> transactionsList = List.from(state.transactions);
       transactionsList[index] = transaction;
@@ -55,14 +60,14 @@ class TransactionCubit extends Cubit<TransactionState> {
 
   //delete transaction from all transactions and objectbox repository
   void deleteTransaction(int index) {
-    //create local copy of transactions
-    List<Transaction> transactionsList = List.from(state.transactions);
-    int id = transactionsList[index].id;
-    transactionsList.removeAt(index);
+    if (index >= 0 && index < state.transactions.length) {
+      //create local copy of transactions
+      List<Transaction> transactionsList = List.from(state.transactions);
+      int id = transactionsList[index].id;
+      transactionsList.removeAt(index);
 
-    transRepository.deleteTransaction(id);
-    emit(state.copyWith(transactions: transactionsList));
+      transRepository.deleteTransaction(id);
+      emit(state.copyWith(transactions: transactionsList));
+    }
   }
 }
-
-
